Update to liblouis 3.0.0.alpha1

+++ src/main/java/org/liblouis/Louis.java
@@ -36,11 +36,11 @@ public class Louis {
 	public interface LouisLibrary extends Library {
 		
 		public int lou_translate(String tableList, WideString inbuf, IntByReference inlen,
-				WideString outbuf, IntByReference outlen, byte[] typeform, byte[] spacing,
+				WideString outbuf, IntByReference outlen, short[] typeform, byte[] spacing,
 				int[] outputPos, int[] inputPos, IntByReference cursorPos, int mode);
 		
 		public int lou_backTranslate(String tableList, WideString inbuf, IntByReference inlen,
-				WideString outbuf, IntByReference outlen, byte[] typeform, byte[] spacing,
+				WideString outbuf, IntByReference outlen, short[] typeform, byte[] spacing,
 				int[] outputPos, int[] inputPos, IntByReference cursorPos, int mode);
 		
 		public int lou_hyphenate(String tableList, WideString inbuf, int inlen, byte[] hyphens, int mode);
